Move to model package.
