+++ Compose-Flight-Search-App/app/src/androidTest/kotlin/com/example/flightsearchapp/testing/model/database/AirportEntitiesTestData.kt
@@ -1,7 +1,7 @@
-package com.example.flightsearchapp.testing.database
+package com.example.flightsearchapp.testing.model.database
 
 import com.example.flightsearchapp.data.database.AirportEntity
-import com.example.flightsearchapp.testing.database.entityextension.testAirportEntity
+import com.example.flightsearchapp.testing.model.database.entityextension.testAirportEntity
 
 val airportEntitiesTestData: List<AirportEntity> = listOf(
     testAirportEntity(1, "ABC"),
